requirements.txt
# pip freeze > requirements.txt
openai==0.27.0
streamlit==1.19.0
chromadb==0.3.6
pdfminer.six==20201018
langchain==0.0.100


